pyproject.toml
[project]
name = "vacation_planner"
version = "0.1.0"
description = "vacation_planner using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "bedrock-agentcore>=1.1.4",
    "boto3>=1.42.26",
    "crewai[tools]==1.8.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]

[project.scripts]
vacation_planner = "vacation_planner.main:run"
run_crew = "vacation_planner.main:run"
train = "vacation_planner.main:train"
replay = "vacation_planner.main:replay"
test = "vacation_planner.main:test"
run_with_trigger = "vacation_planner.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
